Add support for null values in combos
JFace ComboViewers do not like null values in their possible choices,
but null is a legitimate choice for optional references (meaning
"unset"). Replace nulls with a special value (with no risk of conflict
with anything elese) when filling the combo, and convert it back to null
when passing it to the controller if the users chooses it.

Also avoid triggering the controller's updateValue() callback during
refresh when we configure the combo to show the current value.

Change-Id: Ic3d4f53f8c7f5c11a6aee02ef59708024bd63154
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>