[498259] Action buttons broken after switching tab
The action buttons list was cleared when the aboutToBeHidden method was
called which means that when we switched to another tab the action
buttons references where cleared. To fix the issue, the clear action is
just moved to the dispose method.
The issue is fixed for all the widgets able to define action buttons :
list, hyperlink and label.

Bug: 498259
Change-Id: Id6ffc47e392a969747d2561d849bebc10a00333d
Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>