[498748] Fix potential NPE when switching back to an existing tab
Tabs are cached according to their descriptors' getId(), which itself
relies (among other things) on the URI of the semantic EObject. That URI
can change when the object is edited (for example: renaming an EClass in
a .ecore). When switching back to a tab previously created (and cached)
with the now-obsolete URI, descriptorToTab.get(descriptor) will return
null, but the code was not ready to handle that case.

Change-Id: I819a0eef62ae6d8e17dbd262c7bccf37a3080c3b
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>