[496076] Review the tool bar lifecycle
The toolbar actions are not removed when we select another page. Then if
a page without action is selected first the actions are not provided
when a page which defines action is selected. We review the
creation/deletion of the toolbar actions to be done during the
aboutToBeShwon and aboutToBeHidden phases.

Bug: 496076
Change-Id: If849f60e2c5b2929fbe3d1fcfb1d9fb9161d1437
Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>