[530181] Add workarounds for combos under Windows 7
By default, under Windows 7 (and only there), if a combo is opened, it
will ignore concurrent requests to update its content. This means it
can display stale values, but trying to update the displayed values
once the combo is opened is broken under Windows 7.

This is configurable with system property
org.eclipse.eef.avoidUpdatingOpenedCombo if the situations appears in
other systmes we're not yet aware of.

If showing possibly stale values in the combo is an issue, a second
system property, org.eclipse.eef.forceCloseToUpdateCombo, can be set
to force the updating, at the cost of closing the combo. The user has
to re-open it, but at least in this case the content shown is up to
date.

Bug: 530181
Change-Id: Iba6ae5767e208206adbc4207c50894d52dedd0b1
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>